1
1

+++ Assets/01. Scripts/Character/C_StatBaseSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Player Stat", menuName = "SO/Data-Player")]
 public class C_StatBaseSO : ScriptableObject
 {
-    public float maxHp;
+    public int maxHp;
     public float moveSpeed;
     public float critChance;
     public float critMultiplier;
