+++ Assets/01. Scripts/Interface/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damaged(float amount, GameObject attacker = null);
+    void Damaged(int amount, GameObject attacker = null);
 }
